cairos_python_lowlevel/pyproject.toml
[tool.poetry]
name = "cairos_python_lowlevel"
version = "0.9.24"
description = "A client library for accessing Cairos API"
authors = []
readme = "README.md"
packages = [
    {include = "cairos_python_lowlevel"},
]
include = ["CHANGELOG.md", "cairos_python_lowlevel/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
